+++ watergate/src/test/main_lock_table.cpp
@@ -35,11 +35,11 @@ int main(int argc, char *argv[]) {
         t->init(LOCK_TABLE_NAME, resource, LOCK_TABLE_SIZE, true);
 
 
-        uint32_t table_size = t->get_table_size();
+        uint16_t table_size = t->get_table_size();
         lock_table_client *records[table_size];
 
         int count = 0;
-        for (int ii = 0; ii < table_size; ii++) {
+        for (uint16_t ii = 0; ii < table_size; ii++) {
             string app_name("LOCK_TABLE_TEST_");
             app_name.append(to_string(ii));
             pid_t pid = 1000 + ii;
@@ -61,7 +61,7 @@ int main(int argc, char *argv[]) {
 
         LOG_DEBUG("Created [%d] lock records.", count);
 
-        for (int ii = 0; ii < table_size; ii++) {
+        for (uint16_t ii = 0; ii < table_size; ii++) {
             lock_table_client *c = records[ii];
             CHECK_NOT_NULL(c);
             CHECK_AND_FREE(c);
